+++ extension/content.js
@@ -43,3 +43,4 @@ chrome.runtime.onMessage.addListener((message, _sender, sendResponse) => {
 
 initFloatingControl();
 initSelectionTranslation();
+initSubtitleTranslation();